refactor: google analytics
